minor refactor to move class into the actionfurtherevidence package
