+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/HandleEvidenceEventHandlerTest.java
@@ -13,6 +13,7 @@ import org.mockito.Mock;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.actionfurtherevidence.HandleEvidenceEventHandler;
 
 public class HandleEvidenceEventHandlerTest {
